requirements.txt
gradio
google-genai